+++ src/DotNetInstallSdk/GlobalJsonLocator.cs
@@ -1,8 +1,7 @@
 using System.IO;
 using System.Text.Json;
-using System.Text.Json.Serialization;
 
-namespace InstallSdkGlobalTool
+namespace DotNet.InstallSdk
 {
     public class GlobalJsonLocator
     {
