+++ src/DotNetInstallSdk/SdkAcquirer.cs
@@ -1,16 +1,14 @@
 using System;
 using System.Buffers;
+using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Net.Http;
+using System.Security.Cryptography;
 using System.Text.Json;
 using System.Threading.Tasks;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Runtime.InteropServices;
-using System.Security.Cryptography;
 
-namespace InstallSdkGlobalTool
+namespace DotNet.InstallSdk
 {
     public class SdkAcquirer
     {
