+++ tests/DotNetInstallSdk.Tests/TestReleaseData/2.2-releases.json
