+++ tests/DotNetInstallSdk.Tests/TestReleaseData/releases-index.json
