+++ tests/DotNetInstallSdk.Tests/_global.json
